M1-Challenge-Joel-Stockard/Exercise5/CustomerComposition/src/main/java/com/company/Customer.java
package com.company;

public class Customer {
    private String first;
    private String last;
    private String email;
    private String phone;
    private boolean vip;
    private Address shippingAddress;

    private Address billingAddress;


    public Customer(String first, String last, String email, String phone, Address shippingAddress, Address billingAddress, boolean vip) {
        this.first = first;
        this.last = last;
        this.email = email;
        this.phone = phone;
        this.shippingAddress = shippingAddress;
        this.billingAddress = billingAddress;
        this.vip = vip;
    }

    public String getFirst() {
        return first;
    }

    public void setFirst(String first) {
        this.first = first;
    }

    public String getLast() {
        return last;
    }

    public void setLast(String last) {
        this.last = last;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

    public Address getShippingAddress() {
        return shippingAddress;
    }

    public void setShippingAddress(Address shippingAddress) {
        this.shippingAddress = shippingAddress;
    }

    public Address getBillingAddress() {
        return billingAddress;
    }

    public void setBillingAddress(Address billingAddress) {
        this.billingAddress = billingAddress;
    }

    public boolean isVip() {
        return vip;
    }

    public void setVip(boolean vip) {
        this.vip = vip;
    }
}
